Fix: fixed variables processing
